Refactor `TransportType`
